+++ src/settings/FinanceSettings.js
@@ -13,16 +13,19 @@ export default class FinanceSettings extends React.Component {
       component: FundTypeSettings,
       label: <FormattedMessage id="ui-finance.settings.fundTypes.title" />,
       route: 'fund-types',
+      perm: 'ui-finance.settings.view',
     },
     {
       component: ExpenseClassSettings,
       label: <FormattedMessage id="ui-finance.settings.expenseClasses.title" />,
       route: 'expense-classes',
+      perm: 'ui-finance.settings.view',
     },
     {
       component: ExportFundSettings,
       label: <FormattedMessage id="ui-finance.settings.exportFund.title" />,
       route: 'export-fund',
+      perm: 'ui-finance.settings.exportFundAndExpenseClassCodes',
     },
   ];